+++ packages/client-utils/src/main.js
@@ -1,3 +1,4 @@
 export * from './rpc.js';
 export * from './sync-tools.js';
+export * from './vstorage-kit.js';
 export * from './wallet-utils.js';
