+++ packages/client-utils/test/snapshots/exports.test.js.md
@@ -12,10 +12,10 @@ Generated by [AVA](https://avajs.dev).
       'boardSlottingMarshaller',
       'makeAgoricNames',
       'makeFromBoard',
-      'makeRpcUtils',
       'makeStargateClient',
       'makeTendermint34Client',
       'makeVStorage',
+      'makeVstorageKit',
       'makeWalletUtils',
       'pickEndpoint',
       'retryUntilCondition',
